// src/config/enums.ts
export enum ConfigKey {
  PORT = 'PORT',
  NODE_ENV = 'NODE_ENV',
  SDK_FINANCE_BASE_URL = 'SDK_FINANCE_BASE_URL',
  JWT_PUBLIC_KEY_DEV = 'JWT_PUBLIC_KEY_DEV',
  JWT_PUBLIC_KEY_PROD = 'JWT_PUBLIC_KEY_PROD',
  JWT_PRIVATE_KEY_DEV = 'JWT_PRIVATE_KEY_DEV',
  JWT_PRIVATE_KEY_PROD = 'JWT_PRIVATE_KEY_PROD',
  JWT_EXPIRE_TIME = 'JWT_EXPIRE_TIME',
  JWT_REFRESH_EXPIRE_TIME = 'JWT_REFRESH_EXPIRE_TIME',
  REDIS_HOST = 'REDIS_HOST',
  REDIS_PORT = 'REDIS_PORT',

  // Logging Configuration
  LOG_LEVEL = 'LOG_LEVEL',
  LOG_TO_CONSOLE = 'LOG_TO_CONSOLE',
  LOG_TO_FILE = 'LOG_TO_FILE',
  LOG_FILE_PATH = 'LOG_FILE_PATH',
}